+++ templates/estimate_genome_size.sh
@@ -19,17 +19,52 @@ else
                 grep "Estimated genome size:" | \
                 awk '{if($4){printf("%d\n", $4)}} END {if (!NR) print "0"}' > ${OUTPUT}
         fi
-        rm test.msh
+        rm -rf test.msh
+        ESTIMATED_GENOME_SIZE=`head -n1 ${OUTPUT}`
+
+        if [ ${ESTIMATED_GENOME_SIZE} -gt "!{params.max_genome_size}" ]; then
+            # Probably high coverage, try increasing number of kmer copies to 10
+            if [ "!{single_end}" == "false" ]; then
+                mash sketch -o test -k 31 -m 10 -r !{fq[0]} !{fq[1]} 2>&1 | \
+                    grep "Estimated genome size:" | \
+                    awk '{if($4){printf("%d\n", $4)}} END {if (!NR) print "0"}' > ${OUTPUT}
+            else
+                mash sketch -o test -k 31 -m 10 !{fq[0]} 2>&1 | \
+                    grep "Estimated genome size:" | \
+                    awk '{if($4){printf("%d\n", $4)}} END {if (!NR) print "0"}' > ${OUTPUT}
+            fi
+            rm -rf test.msh
+        elif [ ${ESTIMATED_GENOME_SIZE} -lt "!{params.min_genome_size}" ]; then
+            # Probably low coverage, try decreasing the number of kmer copies to 1
+            if [ "!{single_end}" == "false" ]; then
+                mash sketch -o test -k 31 -m 1 -r !{fq[0]} !{fq[1]} 2>&1 | \
+                    grep "Estimated genome size:" | \
+                    awk '{if($4){printf("%d\n", $4)}} END {if (!NR) print "0"}' > ${OUTPUT}
+            else
+                mash sketch -o test -k 31 -m 1 !{fq[0]} 2>&1 | \
+                    grep "Estimated genome size:" | \
+                    awk '{if($4){printf("%d\n", $4)}} END {if (!NR) print "0"}' > ${OUTPUT}
+            fi
+            rm -rf test.msh
+        fi
 
         ESTIMATED_GENOME_SIZE=`head -n1 ${OUTPUT}`
         if [ ${ESTIMATED_GENOME_SIZE} -gt "!{params.max_genome_size}" ]; then
             rm ${OUTPUT}
-            echo "!{sample} estimated genome size (${ESTIMATED_GENOME_SIZE} bp) exceeds the maximum allowed
-                  genome size (!{params.max_genome_size} bp). If this is unexpected, please investigate
-                  !{sample} to determine a cause (e.g. metagenomic, contaminants, etc...).
-                  Otherwise, adjust the --max_genome_size parameter to fit your need. Further
-                  analysis of !{sample} will be discontinued." | \
-            sed 's/^\s*//' > max-genome-size-depth-error.txt
+            echo "!{sample} estimated genome size (${ESTIMATED_GENOME_SIZE} bp) exceeds the maximum
+                  allowed genome size (!{params.max_genome_size} bp). If this is unexpected, please
+                  investigate !{sample} to determine a cause (e.g. metagenomic, contaminants, etc...).
+                  Otherwise, adjust the --max_genome_size parameter to fit your need. Further analysis
+                  of !{sample} will be discontinued." | \
+            sed 's/^\s*//' > !{sample}-genome-size-error.txt
+        elif [ ${ESTIMATED_GENOME_SIZE} -lt "!{params.min_genome_size}" ]; then
+            rm ${OUTPUT}
+            echo "!{sample} estimated genome size (${ESTIMATED_GENOME_SIZE} bp) is less than the minimum
+                  allowed genome size (!{params.min_genome_size} bp). If this is unexpected, please
+                  investigate !{sample} to determine a cause (e.g. metagenomic, contaminants, etc...).
+                  Otherwise, adjust the --min_genome_size parameter to fit your need. Further analysis
+                  of !{sample} will be discontinued." | \
+            sed 's/^\s*//' > !{sample}-genome-size-error.txt
         fi
     elif [ "!{params.genome_size}" == "min" ]; then
         # Use the minimum genome size based on completed RefSeq genomes